graph-pad/requirements.txt
anthropic==0.45.2
faiss_gpu==1.7.2
Flask==3.1.0
google_cloud_aiplatform==1.69.0
habitat_sim==0.2.4.dev20230403
hovsg==1.0.0
imageio==2.35.1
imageio==2.34.0
Imath==0.0.2
kornia==0.7.3
matplotlib==3.7.3
natsort==8.4.0
networkx==3.2.1
numpy==1.23.2
omegaconf==2.3.0
open3d==0.17.0
openai==1.63.2
opencv_python==4.9.0.80
opencv_python_headless==4.8.1.78
OpenEXR==3.3.2
pandas==2.2.3
Pillow==9.5.0
Pillow==11.1.0
protobuf==5.29.3
pypng==0.20220715.0
pyvis==0.3.2
PyYAML==6.0.2
PyYAML==6.0
PyYAML==6.0.2
rerun==1.0.30
scikit_learn==1.6.1
scipy==1.15.2
segment_anything==1.0
setuptools==75.1.0
supervision==0.25.1
tenacity==8.2.3
torch==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.66.5
transformers==4.46.2
ultralytics==8.3.1
vertexai==1.69.0
